feat: release to npm after creating git commit
With `@semantic-release/npm` being used before `@semantic-release/git` the NPM release doesn't contain updated `CHANGELOG.md` and `package.json`.
And that can be very confusing for the consumers of the package.

Simply changing the order of plugins should fix that. GitHub plugin already is in that order, so it will be consistent.